Handle exclusions in PROJECT.scl owned code paths.
For example, given:

```
owned_code_paths = [
    "a",
    "-a/b",
    "a/b/c",
]
```

`a/b/c/d.txt` matches, but `a/b/e.txt` does not.

Implementation: Instead of hand-rolling our own inclusion/exclusion matcher for each file against a set of prefixes, we use a trie for prefix matching.

Unfortunately, there _still_ isn't a PrefixTrie implementation in Guava, so we have to handroll our own implementation. This implementation works specifically with the segments in a `PathFragment`.

PiperOrigin-RevId: 657534818
Change-Id: Ie4e3d48e2ba8d334890d6b83bfaa670c1572e278