`$ bazel test`: always configure `--run_under=//foo` for the exec configuration.
Rationale: test actions run on exec platforms, not the target platform. Configuring a test action for the target platform and running it on the exec platform makes no sense.

This reverts an earlier conversion of `--run_under` from the exec to target platform . The rationale there was:

1. Using the target configuration is consistent with `$ bazel run --run_under`
2. When running tests, bazel uploads both the run_under target and the test itself to the same machine. Therefore they should have the same config.

We should address these points by:

1. Converting `$ bazel run --run_under` similarly(added a TODO)
2. Updating the default test runner similarly

Fixes https://github.com/bazelbuild/bazel/issues/22624.

PiperOrigin-RevId: 657586281
Change-Id: I04e7d712015535caa84824a7cec0b79a663bcfbe