Refine symbolic macro behavior w.r.t. implicitly created input files
This CL prevents automatic InputFile creation for any label whose name is within a symbolic macro's namespace. Previously we only did this when the name exactly matched the symbolic macro's name. This CL also prevents automatic creation when the label is only requested *by* targets within symbolic macros.

For example, if you write srcs = ["foo_bar"] and have a macro named "foo", we will no longer implicitly create an InputFile for "foo_bar". (You can still explicitly declare it with exports_files outside the macro, assuming it doesn't conflict with a target created inside the macro.) In addition, if srcs = ["baz"] is only seen within the body of "foo" but not in the direct evaluation of the BUILD file, we won't create an InputFile for "baz".

The rationale for both of these changes is that it makes input file semantics less entangled with whether or not a macro is evaluated (i.e. concerns for potential future lazy evaluation of macros). Previously, you needed to evaluate macros in order to determine 1) what "dangling references" it contains to undeclared input files, and 2) whether or not someone else's reference to within the macro's namespace is in fact dangling. This CL modifies the semantics to make these questions irrelevant to deciding what InputFiles get automatically declared.

Package.java
- New rulesCreatedInMacros builder field to help track whether or not we should declare input files. Eventually this can be replaced by keeping better track in a Target class of whether or not it was created within a symbolic macro. Note that we're tracking targets declared within a macro, not the opposite, so as to keep the cost low initially. Like ruleLabels, this field is null when not needed, in particular on package deserialization (where implicit input file creation isn't done anyway).
- Update logic in beforeBuild for new behavior, clarify comments.
- Refactor checkDeclaredNameValidForMacro, to pull out a boolean predicate function.

PackageFactoryTest.java
- Expand existing test case.

SymbolicMacroTest.java
- Add a test case for implicit outputs that violate the namespace requirement. Not directly relevant, but I had this lying around in unsubmitted code for a while.
- Add test that macro can refer to input files outside of it, whether or not they're implicitly created
- Add test that macro cannot itself cause the implicit creation of an input file.
- Normalize some text blocks to match the rest of the file.

Work toward #19922.

PiperOrigin-RevId: 658856368
Change-Id: I7ca5b0692ecf7a7695494d14b869e699eb4bafb5