Blaze/Bazel changes for Java 17 Record desugaring
createDexMergerActions continues to be exposed to Starlark due to it's use for SpawnActionTemplate which is not exposed to Starlark. We must pass a file containing global information to the final dexmerging step in order to support Java 17 Record desugaring. This change adds an optional parameter to the Starlark interface to forward the globals file to the Dex Merger.

PiperOrigin-RevId: 660045021
Change-Id: Ie3b961ff2687f48ca29a94e67180c0e5875ddbe3