Properly flush remaining buffered output at the end of the build.
Previously, this didn't happen in all cases, in particular if `aquery` output
included a newline character, e.g. if the length of a string is 10 bytes it
would be cut off.

A better, but much more involved fix would be to circumvent `UiEventHandler` for `{,a,c}query` output to stdout.

PiperOrigin-RevId: 660332325
Change-Id: I1a816028660da6598a001740bd7e3c95f9a1a098