Index symbolic macros by id rather than name
This is to prep for allowing multiple macros by the same name, in the case where one is a submacro of another.

Same-named macros are only allowed when one is nested inside the other. This nesting can be transitive, but it must be contiguous, e.g. not foo -> bar -> foo; this is enforced by macro naming rules.

Allowing same-named macros is important for ensuring that, when a rule is migrated to a symbolic macro, a caller who was using the rule as their main target does not get broken by the new name conflict between macros.

The id is formed by concatenating the name with ":1", ":2", etc., starting with the outermost macro of a given name and incrementing for each nested macro. The id is not intended to be stable across changes to a build, but it is deterministic and ensures uniqueness, making it suitable for map keys and in APIs like blaze query.

The common case is that almost all macros have a name ending in ":1". We could simplify down to just the name in that case, and might still do so for user interfaces like blaze query. But internally it's best to key strictly on the id, not the name, to avoid errors that are only surfaced in the relatively rare case of an id other than ":1".

MacroInstance.java:
- Track the depth of a macro in the chain of all macros having the same name.

MacroClass.java:
- Compute the depth number based on the current stack of running macros.
- Rename serialization-related class ("MacroId" -> "MacroClassId") to avoid confusion with the concept of a macro instance's id.

Package.java
- Index macros map by id rather than name.
- Switch macro stack to use a separate frame class, which will be handy for subsequent CLs adding naming restrictions.
- Add helper methods for retrieving macro stack info.
- Update comments.

Work toward #19922.

PiperOrigin-RevId: 660360402
Change-Id: Ie531474ce4736f803a2fdc7527e51fc6ac704b0f