Symbolic macros and targets cannot share names (except for main targets)
It is now prohibited to define a symbolic macro and a target in the same package that share the same name, except that a target may have the same name as its ancestor macros (a macro's "main target").

This helps avoid breakages when refactoring a rule into a symbolic macro, due to creating a conflict between two macros that was previously allowed as a conflict between a macro and a target.

Package.java
- Refactor checkForExistingMacroName() to take a name rather than a macro instance, and call it for target names as well.
- Conversely, have checkMacroName() check for target names as well. (I.e., we need to update both methods to cover both possible orders for which thing gets declared first.)
- Update some comments. In the check...() methods, simplify/steamline javadoc.

PackageFactory.java
- Invert a bunch of test cases.

SymbolicMacroTest.java
- Remove a now illegal target/macro conflict from a test case.
- Add test cases for conflicts among two main siblings of a macro.

Work toward #19922.

PiperOrigin-RevId: 660865483
Change-Id: Ic1d3a90faab959ead8d103599b79fe67ff733192