[7.3.1] Fix expected lockfile version to change based on `--incompatible_use_plus_in_repo_names`
We write a different lockfile version based on `--incompatible_use_plus_in_repo_names` (see HACK in `BazelLockFileModule.java`). But we don't _expect_ a different lockfile version based on that flag; this causes us to consider the lockfile _always_ out-of-date when that flag is set.

Fixes https://github.com/bazelbuild/bazel/issues/23279.