Avoid pointless exception wrapping.
SpawnRunner#exec is already allowed to return a ForbiddenActionInputException or IOException, which are converted into an ExecException in AbstractSpawnStrategy#exec. The wrapping provides no useful context, as this error occurs during input prefetching, which is independent from the spawn runner.

(Note: the PREFETCH_FAILURE code will be replaced with a generic EXEC_IO_EXCEPTION, but the former is only used by the WorkerSpawnRunner, which is weird anyway. If we cared sufficiently about the distinction, it would make more sense to throw an EnvironmentalExecException from prefetchInputsAndWait, regardless of the runner being used.)

PiperOrigin-RevId: 662839172
Change-Id: Ic04f8907931a60bd6cc9130340c4a6ef4d3f9c80