Ban recursion in symbolic macro expansions
Recursive instantiation of a symbolic macro is now prohibited. Recursive instantiation is when a macro instance of macro class A either directly or indirectly declares another instance of A. This prohibition applies whether or not macros are evaluated eagerly on the Java call stack.

The motivation for banning recursion is mainly symmetry with ordinary Starlark evaluation of .bzl files, and the fear that leaving this open will encourage users to abuse symbolic macros simply to do an end-run around the existing recursion check.

- MacroInstance now stores the parent MacroInstance it was declared within (null for top-level macros).
- MacroClass now stores the label of the .bzl where it was exported, in addition to its export name. This is useful for tracebacks but also could be used for an equals() behavior that's not reference equality.
- Add recursion check to MacroClass#executeMacroImplementation. A violation is treated similarly to a Starlark evaluation error. The error message includes the package, instance name, and macro class of each macro that led to the cycle, printed outermost first to be in the style of a Starlark/Python traceback.
- Updated PackageFactoryTest, replacing the recursive tree macro with a simpler linear one now that there's no interesting legal recursive behavior to test. Also add a new non-recursive test for transitive deferred macro evaluation, to replace the recursive one that is now illegal.

Work toward #19922.

PiperOrigin-RevId: 663746520
Change-Id: I5469b598fd0a7ccf844ec2271e338ff97daf1971