Allow ':' in target and filenames if not the first character
Especially in the Perl community, filenames containing `:` are common. Labels can still be parsed unambiguously if target names can contain colons by declaring that the first colon in a given label string is interpreted as the separator between package and target name. This only requires prefixing with `:` when referring to targets containing colons in the same package.

Target and filenames starting with `:` are still not allowed to prevent confusion with references to targets.

H