For symbolic macros, implicitly append the current location to target's visibility attr
The design of Macro-Aware Visibility says that a target is visible to the locations given in its visibility attribute, plus the location where the target is declared (its BUILD file, or the package of the .bzl where the current symbolic macro was exported).

This CL automatically concatenates the declaration location into the materialized visibility attribute of targets that are declared inside symbolic macros. Since we're munging the visibility attribute, this is a loading-time effect, and the result can be seen in `bazel query` output (or in our unit tests, with Target#getVisibility). When we add a visibility attribute to symbolic macros, they will have the same behavior, with the result that the visibility argument seen in macro implementation functions will include the caller's location.

Targets that are not declared within a symbolic macro are unaffected, and will continue to show visibility attribute values that do not necessarily include their package explicitly. A subsequent change will work around this behavioral difference between targets in macros and out of macros, by adding the declaration location for the latter at analysis time, in the visibility provider. We went with this hybrid approach in order to avoid a massive and immediate change to all targets' visibilities. It's not clear whether we need to change this in the future for consistency.

Changes:
- Add a map in Package, from target name to its innermost declaring macro. Add a couple TODOs for future optimization. Add a corresponding utility function Target#getDeclaringMacro.
- Add Package#visibilityWithCurrentMacro(), a utility function for munging a visibility attribute value to append the location of the currently executing symbolic macro's definition. Apply this to rule declarations in RuleFactory, and also to exports_files in StarlarkNativeModule.
- Clean up RuleFactory#generatorAttributesForMacros by moving attribute map reconstruction into the caller where it can be handled alongside visibility. Clarify comments.
- Add TODO comment about reserved macro attr names.
- Add visibility attr tests to PackageFactoryTest. Also rename a helper method and add another helper method for unrelated symbolic macro tests.

Work toward #19922.

PiperOrigin-RevId: 665419955
Change-Id: I98ae1acdfbed3662ae00624d663dac1d35b9a8ec