SkyQueryEnvironment does not log query transformation for huge queries.
The `SkyQueryEnvironment` features a query transformation step that can optimize queries for more efficient execution. The before-and-after queries are logged to support retrospective debugging of the query transformation step. This logging stops being useful when queries are far too large for inspection in human-readable form. Skipping this logging step - without stringify-ing the QueryExpression in either its 'before' or 'after' forms - can avoid allocating Strings that are hundreds of megabytes or gigabytes in size.

RELNOTES: n/a
PiperOrigin-RevId: 666355603
Change-Id: Idf3acc464210c52af06ce1828bb0300180a71d61