Ensure the visibility provider can accurately track targets' declaration locations
This CL adds an accessor to VisibilityProvider that tells whether the target was instantiated within at least one symbolic macro. This is used in a follow-up CL to help implement Macro-Aware Visibility, without breaking the logic that a target declared outside a symbolic macro is visible to its own package regardless of what its visibility attribute says. (In addition, under Google's own internal visibility policy, we preserve that a target in //java/foo that's not in a symbolic macro is visible to //javatests/foo.)

To avoid storing an extra bool in every RuleConfiguredTarget, I implemented this with the help of a singleton marker provider. This is of course more expensive than a bool, but it's a cost that's only paid for targets within symbolic macros. Other configured target types just get a new bool field where needed, e.g. in InputFileConfiguredTarget but not in OutputFileConfiguredTarget.

Also added lots of comments, javadoc, and tests.

A previous draft of this CL also added the current package to the VisibilityProvider of targets not declared in symbolic macros, to work around the fact that these are omitted from said target's visibility attribute. But this turned out to be a memory regression and not necessary anyway, since the information can be reconstructed at visibility checking time from the bool accessor. If we needed to add this in the future anyway, it could probably be done more efficiently by interning the added visibility information, taking advantage of the fact that most targets in a package share the same visibility.

Notable changes:
- ConfiguredTargetFactory: Factor public/private visibility constant values into VisibilityProvider. Add explanatory comment to convertVisibility.
- VisibilityProvider: Add isCreatedInSymbolicMacro() accessor. Add explanatory comments.
- AbstractConfiguredTarget: Delete convenience constructor. It just made the edge cases that needed it more confusing to read.
- InputFileConfiguredTarget: Store an extra bool field to implement isCreatedInSymbolicMacro(). (We get away with not storing this field for output files.)
- PackageGroupConfiguredTarget: Delete unused alternate constructor.
- RuleConfiguredTarget: Materialize the isCreatedInSymbolicMacro bit as whether or not the singleton instance of a new private provider class is present on the target.
- IncompatibleTargetChecker: When creating RuleConfiguredTargets here, just assume they're not in a symbolic macro. My sense is that this is in line with the existing behavior of conservatively assuming targets are public for the purpose of compatibility checking.

Work toward #19922.

PiperOrigin-RevId: 666380629
Change-Id: Iac15bf3612a1d2144e0fcd2d1c49c1496ae19466