[7.4.0] Fix in-process materialization of unresolved symlinks in runfiles (#23391)
Unresolved symlinks, when consumed as runfiles with
`--experimental_inprocess_symlink_creation`, are now created as symlinks
directly pointing to their target path instead of to the absolute path
of the unresolved symlink artifact under the execroot, just like with
out-of-process symlink creation.

Along the way, fix the propagation of
`--experimental_inprocess_symlink_creation` to the exec configuration,
which previously would have caused the new test cases to fail open.

Fixes #23327

Closes #23328.

PiperOrigin-RevId: 665754549
Change-Id: I6b240ca27dac513738056b015c97baec91677c4e

Closes #23329