[7.4.0] Add path mapping support for C++ compile action templates (#23349)
C++ action templates are now properly path mapped.

Since the default Unix toolchain doesn't support object file groups on
macOS, `apple_support` is needed in the test. This requires wiring up
`LocalEnvProvider` in the remote worker to get `DEVELOPER_DIR` to be
set, which in turn requires improving the runfiles handling for the the
worker.

Work towards #6526

Closes #22890.

PiperOrigin-RevId: 657733074
Change-Id: I132e338d7a44964a8a1aad3062a5c9a4c8e79e57

Commit
https://github.com/bazelbuild/bazel/commit/23f3be03262c89e7c2d18a110c187ee5c760af31

Co-authored-by: Googler <ilist@google.com>