[7.4.0] Upgrade to Bazel 7.3.1 (#23475)
I was originally going to cherry-pick
https://github.com/bazelbuild/bazel/commit/24fcc3e13d6f67ffe8890e359b1216900f8a328c,
but then realized that it would require a lot of "reverse hacks" if we
set `--incompatible_use_plus_in_repo_names` -- because this time,
"builder bazel" uses plus, while "bazel being built" uses tilde. So it's
actually simpler if we don't turn on the flag for release-7.4.0, which I
think is fine.