Implement the bulk of Macro-Aware Visibility
This modifies the visibility check to test the consuming target's declaration location, rather than the package it lives in.

The change to production code is deceivingly small because most of the work was done in the two prerequisite CLs unknown commit and unknown commit, which ensure the visibility provider permits access to the place where the dependency was declared.

The isSameLogicalPackage() check is also gated to only apply when the dependency target is not created in a symbolic macro. Within a macro, the visibility attribute is authoritative and no implicit assumptions are needed. (Though a future change needs to help propagate the same-logical-package down through a macro's `visibility` parameter.)

Added half a kLOC of test cases. Also gave BuildViewTestCase the ability to *not* patch in --default_visibility=public.

PiperOrigin-RevId: 669978176
Change-Id: I641cc96c3a5c46323ca4304ac0c69bb65935be42