Reduce verbosity of BulkTransferException
BulkTransferException was sometimes overly verbose in the console,
particularly when many transfers failed for the same reason.

This commit shortens and normalizes BulkTransferException messages
to improve the efficiency of existing event deduplication, such as
that performed by RemoteExecutionService.report().

Normalization involves not mentioning individual digests in exception
messages from ByteStreamUploader. Instead, we now wrap
StatusRuntimeException directly in an IOException, which is a common
pattern in GrpcCacheClient.