Do not zip undeclared test outputs by default.
Zipping was originally introduced as a workaround for the lack of support for directory outputs (aka tree artifacts). It creates unnecessary work and introduces a dependency on an external tool.

RELNOTES[INC]: `--zip_undeclared_test_outputs` now defaults to false, causing undeclared test outputs (i.e., files written to `$TEST_UNDECLARED_OUTPUTS_DIR` by a test) to be produced as a directory instead of a zip file.

PiperOrigin-RevId: 671305051
Change-Id: I070c2f8de8e1dcf204c8e0ffa1326afc7ac20219