Set --windows_enable_runfiles for runfiles_test.sh on Windows.
This test asserts the presence of symlinks in runfiles trees, but it currently works without --windows_enable_runfiles because symlinks are created by an external process, which will unconditionally attempt to create them. Once --experimental_inprocess_symlink_creation is flipped, the --windows_enable_runfiles flag will be respected (i.e., symlinks will be created only if the flag is enabled, otherwise falling back to a copy, as is already the case outside of runfiles).

PiperOrigin-RevId: 671342054
Change-Id: I0a5b1b8ea2c83a0c831ea41790cc3f2e458abda1