Flip --experimental_inprocess_symlink_creation.
RELNOTES: Symlink trees are now created through direct filesystem calls by default, instead of delegated to a helper process. On Windows, this entails respecting the `--windows_enable_symlinks` flag, falling back to a copy when the flag is unset (the helper process always attempts to create symlinks, irrespective of the flag). Set `--noexperimental_inprocess_symlink_creation` to temporarily revert to the previous behavior, which will be removed in a future release.
PiperOrigin-RevId: 671352606
Change-Id: Ie353a33bda4e7b642428bc2a77d733422a799086