Don't create-and-retain a `PathFragment` instance for `suffix` for each `Runfiles` instance.
In practice, `suffix` is always the workspace name string, so all these `PathFragment`s are wasteful dupes. Instead, we create the `PathFragment` on the fly (and then throw it away) when needed. Thus this is essentially a logical rollback of unknown commit (in other words, that CL was actually a memory regression).

As an alternative, I considered threading through a logical singleton `PathFragment` to all `Runfiles` instances (e.g. from `RuleClassProvider`) but that had a much larger blast radius.

While I'm here, rename `suffix` to `prefix`. Afaict, the code actually means to say "prefix" and "suffix" was a mistake in unknown commit.

PiperOrigin-RevId: 671429745
Change-Id: I1e72588859ceab39bcd3f05fa8760aa0d7096b79