Report CPU and wall times in profiler output for build commands again.
This was broken in https://github.com/bazelbuild/bazel/commit/cb42a1dd08ab984943c1b85a5ae83dd1afd2d564. The reason is that MetricsCollector assumed that the profiler was active while calling finishTimingMetrics(), which used to be the case before that change. However, that change made finishTimingMetrics() be called in CommandPrecompleteEvent and (somewhat confusingly) BuildSummaryStatsModule shuts down the profiler in BuildCompleteEvent, which happens before.

Ideally, we'd shut down the profiler at the last minute, but we can't do that because we publish the profiler data to the BEP, which in turn shuts down in BuildCompleteEvent. So we must shut down the profiler in BuildCompleteEvent, but that doesn't happen on non-build commands so we must also shut it down some place that is called on every command. This of course means that we lose profiler events between BuildCompleteEvent and the actual end of the command.

The profiler is changed so that timing data is available after .stop(), which makes MetricsCollector work again.

RELNOTES: None.
PiperOrigin-RevId: 671727112
Change-Id: Ibf336a7a387449a3edbaffb3d399e5b5e9ed0ff4