Implement combined cache in `RemoteCache`.
Previously, combined cache is implemented by `DiskAndRemoteCacheCliet` which is hidden behind the interface `RemoteCacheClient`. This introduced a lot of issues because disk cache is really not a variant of remote cache and we need to handle them differently.

This CL moves the logic from `DiskAndRemoteCacheCliet` to `RemoteCache` so that the code is aware of two caches and:

1. `casUploadCache` is only applied to remote cache.
2. Every functions in `RemoteExecutionCache` only operates remote cache.

The class `RemoteCache` should really be renamed to something else because it doesn't only have remote cache. But this should be a follow-up CL.

Fixes #20962.
Closes #21213.

It's also releated to #20296, but I would like the add the mentioned integration test before close it.

PiperOrigin-RevId: 671740437
Change-Id: Ia55ef879d313d111874b4f8376f66a7d1fceebf7