Fully parse options in `ParsedFlagsFunction`.
Currently, `ParsedFlagsValue` stores a list of native options as strings and parsing is done on demand. Instead, do the parsing during the evaluation of `ParsedFlagsFunction` and store an `OptionsParsingResult` in `ParsedFlagsValue`. This moves caching of options parsing to Skyframe, meaning there's one less need for `BuildConfigurationKeyCache`, which we are aiming to remove. Similarly, `OptionsParsingException` for native option parsing errors is now communicated via Skyframe error handling.

`mergeWith()` is moved from `NativeAndStarlarkFlags` to `ParsedFlagsValue` so that it can make use of the already-computed `OptionsParsingResult`. `NativeAndStarlarkFlags` is kept around mostly because `OptionsParsingResult` doesn't implement equality and is not serializable (though `ParsedFlagsValue` is never serialized in practice currently).

PiperOrigin-RevId: 672520782
Change-Id: Ibb528cdb5b34129b752271f570ad598196d9e79b