Add StarlarkMaterializingLateBoundDefault.
It is now possible to add a materializer= argument to attr.label() and attr.label_list() and the Starlark function there is called as it should be, although by a mechanism that's very different from the eventual one and it's more like an artifact of the exact way materializers are implemented than by design. It doesn't get any arguments, though, so it can't do anything useful yet.

RELNOTES:
PiperOrigin-RevId: 672933394
Change-Id: I93c01c5fda89a2fa240b49f82353a2560f01e9cc