Remove test_only_target from cc_proto_library
It looks weird that the C++ proto aspect is over proto_library with testonly set and even weirder over _test, because that isn't a proto_library.
Removing it to reduce complexity and because test_only_target is protected APIs.

The effect of setting this bit is, that under a certain set of flags, FDO indexing is prevented on tests.

PiperOrigin-RevId: 673381037
Change-Id: I5e8ce0389fe01d56fc4f1ae39a5a25aabfcaa846