Remove `BuildConfigurationKeyCache`.
`BuildConfigurationKeyCache` leads to correctness issues because it results in missing skyframe dependency paths from configured targets to source files, such as the `platform_mappings` file or a `BUILD` file defining a `platform` rule.

It is expected that previous changes offset the additional cpu cost of this change. It should add a small amount of additional memory from additional skyframe edges.

PiperOrigin-RevId: 673553258
Change-Id: I3daf7c6be1e4df4507abab38f32abc1051f5b8ab