[7.4.0] Convert values to Starlark values when concating string_list_dict type (#23603)
Convert values to Starlark values when concating string_list_dict type.

This is because Dict checks that all the values put in it are Starlark
valid values. But string_list_dict value is of type List<String>, when
Java's List isn't a Starlark value.

Fixes #23065

My first PR to bazel :star_struck: Would love you to review the PR.
 Checked that the test didn't pass before the change and did after it.

Closes #23424.

PiperOrigin-RevId: 673463108
Change-Id: Ib2cdce7d94243f4e3bda23203a196fee4e766189

Commit
https://github.com/bazelbuild/bazel/commit/f1f8d5829a987854d49913058087765db68fc92c

---------

Co-authored-by: Lior Gorelik <lior10rrr@gmail.com>
Co-authored-by: Alexandre Rostovtsev <arostovtsev@google.com>