Remove additional_exported_hdrs hack from cc_proto_library
It's weird. The files are not generated on this targets, yet C++ proto_library claims to have them.
Those headers should already be provided by runtime, which is linked into every such proto_library.

Removing it also, because this is the only use of it on cc_common.compile. The only way where passing
headers is possible as strings (and not files).

Nothing fails with this removed (tested by running it against Bazel unit tests and against tests in protobuf repo).

PiperOrigin-RevId: 673675704
Change-Id: I976ceaf2101010c7da6c89adbaa97c44fe0323a7