Implement a `RemoteAnalysisCachingDependenciesProvider` interface.
This interface groups together the functionalities needed for serialization, such as `ObjectCodecs`, `FingerprintValueStore`, and the `PathFragmentPrefixTrie` (for active_directories boundary matching).

The implementation is instantiated immediately after evaluating `PROJECT.scl`.
The implementation can then be passed as an argument into the analysis and
execution phases for SkyFunction in-line deserialization operations.

For this CL, we are just updating `FrontierSerializer` to use this provider
instead of individual dependency objects.

PiperOrigin-RevId: 673722701
Change-Id: I7c0426185d2dba9659d204019f022e5bb5a77725