Implement legacy_symbols
In the repositories that don't have autoloads we also expose native.legacy_symbols.
Those can be used to fallback to the native symbol, whenever it's still available in Bazel.

Fallback using a top-level symbol doesn't work, because BzlCompileFunction would throw an error when it's mentioned.

legacy_symbols aren't available when autoloads are not enabled.

The feature is intended to be used with bazel_features repository, which can correctly report native symbols on all versions of Bazel.

PiperOrigin-RevId: 673741927
Change-Id: I2334030d70cbb944b92784e32a184841ea238d51