Distinguish between C++ tools to copy and tools to link in mock configuration.
Also changes copyTool to overwrite target file. This is important
because there may be a difference between a version contained in
builtins and the one being copied in (copies of BZL files in builtins
are sometimes "stripped" down), and this matches the behaviour of
"linkTool" which always removes existing files.

PiperOrigin-RevId: 673751751
Change-Id: I550397f4daa81e48b99c761eeffd22cbe770bbda