Update RegexFilterConverter to reject values starting with `--`.
This happens when a user does not pass a value to a regex-valued flag.

For example, in this command:
```
bazel build --toolchain_resolution_debug --compilation_mode=opt -- //target
```

The value of the `--toolchain_resolution_debug` flag is
`--compilation_mode=opt`, which means "ignore any toolchain type or target
whose label looks like '-compilation_mode=opt'".

This usage is now an error with an explanatory message. If the use is expected,
then the dashes can be escaped with backslashes.

PiperOrigin-RevId: 673953388
Change-Id: I7f8fc81365c75023b0fccefaeb29f468b2f14d2a