Use RuleClass rather than StarlarkRuleFunction as the entry point for RuleInfoExtractor
Note that this requires moving the Starlark docstring from StarlarkRuleFunction to
RuleClass, and keeping a copy of the Starlark file label (which may be null in
the legacy unexported repo rule case) in the RuleClass too.

This also means RuleInfoExtractor now supports native rules - although it cannot (yet)
extract docs from non-Starlark-defined attributes.

Prerequisite for allowing query and packageloader to introspect a rule target's rule class.

PiperOrigin-RevId: 673977421
Change-Id: I926c0997baa4b2f8faec28597040a7b77e5f1b9f