Enable CC toolchain to use in-tree builtin includes
If the toolchain is composed in our build tree and we use an
internal (relative) path for the `builtin_sysroot` that we pass to
`create_cc_toolchain_config_info`, and we pass a value like
`%sysroot%/usr/include` in `cxx_builtin_include_directories` then we
will fail with missing dependency declarations.

Instead of only using the absolute paths for builtin includes, we can
use all of them and do an additional matching check after we
relativized the paths to the individual include items.