GrpcRemoteDownloader: optionally propagate credentials to remote server
In a multi-tenancy server deployment setup, the clients might want to
treat the remote downloader server as a pull-through proxy and use it to
download from private storage systems.

Currently, we do support it via --remote_downloader_headers. However
this scheme does not apply to the specific URL, while credentials and
authentication could sometimes be host/domain specific.

Add a flag to let users opt-in to credentials propagation to the remote
server. This is off by default as not all remote servers can be
trusted. When the flag is enabled, URL-specific credentials from Netrc
or a custom credentials helper can be propagated to the remote server.

The server implementation needs to support the new
`http_header_url:<url-index>:<header-key>` qualifier where the
`url-index` is a 0-based position of the URL inside the
FetchBlobRequest's uris field. This new qualifier is modeled after the
existing `http_header` qualifier.

Fixes #23499

Closes #23578.

PiperOrigin-RevId: 674388422
Change-Id: Iaa2f5dd0bdffd9385d8f229458810442c8ca3ddc