Disable git fsmonitor in git_worker to fix hangs when fetching repositories
This commit adds the `-c core.fsmonitor=false` arguments to all git commands (excluding `git init`) in `git_worker.bzl`.
The purpose of this is to disable git from spawning a file system monitor, when many of these are spawned on a machine that is already under load or out of file descriptors, the git commands will hang indefinitely.

Fixes https://github.com/bazelbuild/bazel/issues/21438